combine OfferPage into useOfferPage
